requirements.txt
Django==5.2.4
djangorestframework
django-filter
drf-spectacular
pytest
pytest-django
pre-commit
black
isort
flake8
